Move getLoaders function to its own module
